+++ build.rs
@@ -1,6 +1,10 @@
 fn main() {
-    #[cfg(all(not(feature = "libfuse"), not(target_os = "linux")))]
-    unimplemented!("Building without libfuse is only supported on Linux");
+    #[cfg(all(
+        not(feature = "libfuse"),
+        not(target_os = "linux"),
+        not(target_os = "macos"),
+    ))]
+    unimplemented!("Building without libfuse is only supported on Linux or Macos");
 
     #[cfg(feature = "libfuse")]
     {