+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.10.0 - UNRELEASED
 * Improve error messages
+* Add support for building without libfuse on Macos
 
 ## 0.9.1 - 2021-09-07
 * `forget` and `batch_forget` no longer require that `AllowRoot` be set